feat(cookies-authentication): setting __session cookies for firebase-admin
